Broke something, fixed it
MS-AutoQC is meant to be used as an offline web app. The idea is that as an instrument run is live, the data will be copied over to bufferbox2 (a SQL database hosted on Stanford servers) and then copied over to Google Drive by Sarah's script (effectively bypassing the need for a Stanford VPN). So, MS-AutoQC should re-download the files every time a user selects a study, so as to "refresh" the files as more data is acquired.